pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arcanite"
version = "0.1.0"
description = "A tarot reading engine with deterministic and LLM-powered interpretations"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Kate Louie" }
]
keywords = ["tarot", "divination", "oracle", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.18",
    "openai>=1.0",
    "httpx>=0.27",
]
pdf = [
    "typst>=0.11",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]
all = [
    "arcanite[llm,pdf,dev]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/arcanite"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
